edited the organizations page layout
